fix wrong model field data type

+++ app/src/main/java/com/jokosupriyanto/filmapp/data/model/Rating.kt
@@ -1,5 +1,5 @@
 package com.jokosupriyanto.filmapp.data.model
 
 data class Rating(
-    val average: Int
+    val average: Double
 ) { }
